Don't load the pyramid application when loading the model
