+++ c2cgeoportal/tests/test_mapserverproxy_route_predicate.py
@@ -32,7 +32,7 @@ from unittest import TestCase
 from pyramid.request import Request
 from pyramid.threadlocal import get_current_registry
 
-from c2cgeoportal import MapserverproxyRoutePredicate
+from c2cgeoportal.pyramid_ import MapserverproxyRoutePredicate
 
 
 class TestMapserverproxyRoutePredicate(TestCase):
