+++ doc/integrator/authentication.rst
@@ -41,7 +41,7 @@ main ``__init__.py`` file, and set ``authentication_policy`` to a
 
    from pyramid.config import Configurator
    from pyramid.authentication import RemoteUserAuthenticationPolicy
-   from c2cgeoportal import locale_negotiator
+   from c2cgeoportal.pyramid_ import locale_negotiator
    from c2cgeoportal.resources import FAModels, defaultgroupsfinder
    from ${package}.resources import Root
 
@@ -208,7 +208,7 @@ LDAP would look like this:
 .. code:: python
 
     def custom_user_validator(request, username, password):
-        from c2cgeoportal import default_user_validator
+        from c2cgeoportal.pyramid_ import default_user_validator
         from pyramid_ldap import get_ldap_connector
         connector = get_ldap_connector(request)
         data = connector.authenticate(username, password)
@@ -228,7 +228,7 @@ Full example using pyramid_ldap, see `# LDAP` / `# END LDAP` blocs.
     from pyramid.config import Configurator
     # LDAP
     # get_user_from_request also needed for the same reason
-    from c2cgeoportal import locale_negotiator, add_interface, \
+    from c2cgeoportal.pyramid_ import locale_negotiator, add_interface, \
         INTERFACE_TYPE_NGEO, get_user_from_request
     # END LDAP
     from c2cgeoportal.lib.authentication import create_authentication
@@ -247,7 +247,7 @@ Full example using pyramid_ldap, see `# LDAP` / `# END LDAP` blocs.
     # LDAP
     # authenticate on LDAP and return cleaned user data
     def custom_user_validator(request, username, password):
-        from c2cgeoportal import default_user_validator
+        from c2cgeoportal.pyramid_ import default_user_validator
         from pyramid_ldap import get_ldap_connector
         connector = get_ldap_connector(request)
         data = connector.authenticate(username, password)
